# Conflicts:
#	base/BUILD.gn
#	base/logging.cc
#	base/mac/launch_application.mm
#	base/synchronization/waitable_event_apple.cc
#	build/config/compiler/BUILD.gn
#	chrome/app/chrome_exe_main_mac.cc
#	chrome/browser/focus_ring_browsertest.cc
#	chrome/browser/ui/views/frame/browser_non_client_frame_view_mac.mm
#	chrome/services/mac_notifications/BUILD.gn
#	chrome/services/mac_notifications/notification_test_utils_mac.h
#	components/remote_cocoa/app_shim/native_widget_mac_nswindow.mm
#	content/app_shim_remote_cocoa/web_drag_source_mac.mm
#	content/browser/accessibility/browser_accessibility_state_impl_mac.mm
#	content/browser/web_contents/web_drag_dest_mac_unittest.mm
#	device/bluetooth/bluetooth_low_energy_device_mac.mm
#	device/fido/BUILD.gn
#	net/base/platform_mime_util_mac.mm
#	sandbox/policy/mac/audio.sb
#	ui/accelerated_widget_mac/ca_layer_tree_unittest_mac.mm
#	ui/base/BUILD.gn
#	ui/base/clipboard/clipboard_ios.mm
#	ui/color/mac/native_color_mixers_mac.mm
#	ui/gfx/animation/animation_mac.mm
#	ui/gfx/font_fallback_mac.mm
#	ui/native_theme/native_theme_mac.mm
#	ui/views/test/widget_test_mac.mm

+++ chrome/browser/ash/app_list/search/local_image_search/annotation_storage.cc
@@ -5,25 +5,23 @@
 #include "chrome/browser/ash/app_list/search/local_image_search/annotation_storage.h"
 
 #include <algorithm>
-#include <iterator>
-#include <map>
-#include <string>
 
 #include "base/logging.h"
-#include "base/memory/scoped_refptr.h"
 #include "base/strings/strcat.h"
 #include "base/strings/utf_string_conversions.h"
-#include "base/task/sequenced_task_runner.h"
-#include "base/task/thread_pool.h"
 #include "chrome/browser/ash/app_list/search/local_image_search/image_annotation_worker.h"
+#include "chrome/browser/ash/app_list/search/local_image_search/search_utils.h"
 #include "chrome/browser/ash/app_list/search/local_image_search/sql_database.h"
 #include "chromeos/ash/components/string_matching/fuzzy_tokenized_string_match.h"
-#include "chromeos/ash/components/string_matching/tokenized_string.h"
-#include "sql/database.h"
 #include "sql/statement.h"
 
 namespace app_list {
 namespace {
+
+using FuzzyTokenizedStringMatch =
+    ::ash::string_matching::FuzzyTokenizedStringMatch;
+using TokenizedString = ::ash::string_matching::TokenizedString;
+using Mode = ::ash::string_matching::TokenizedString::Mode;
 
 constexpr double kRelevanceThreshold = 0.79;
 constexpr int kVersionNumber = 3;
@@ -34,6 +32,9 @@ constexpr int kVersionNumber = 3;
 // The table cannot exist when calling this function.
 int CreateNewSchema(SqlDatabase* db) {
   DVLOG(1) << "Making a table";
+  if (!db) {
+    return 0;
+  }
 
   static constexpr char kQuery[] =
       // clang-format off
@@ -43,24 +44,27 @@ int CreateNewSchema(SqlDatabase* db) {
           "last_modified_time INTEGER NOT NULL,"
           "is_ignored INTEGER NOT NULL)";
   // clang-format on
-  sql::Statement statement = db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  if (!statement.Run()) {
+  std::unique_ptr<sql::Statement> statement =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+  if (!statement || !statement->Run()) {
     return 0;
   }
 
   static constexpr char kQuery1[] =
       "CREATE INDEX ind_annotations_label ON annotations(label)";
 
-  sql::Statement statement1 = db->GetStatementForQuery(SQL_FROM_HERE, kQuery1);
-  if (!statement1.Run()) {
+  std::unique_ptr<sql::Statement> statement1 =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery1);
+  if (!statement1 || !statement1->Run()) {
     return 0;
   }
 
   static constexpr char kQuery2[] =
       "CREATE INDEX ind_annotations_image_path ON annotations(image_path)";
 
-  sql::Statement statement2 = db->GetStatementForQuery(SQL_FROM_HERE, kQuery2);
-  if (!statement2.Run()) {
+  std::unique_ptr<sql::Statement> statement2 =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery2);
+  if (!statement2 || !statement2->Run()) {
     return 0;
   }
 
@@ -68,17 +72,47 @@ int CreateNewSchema(SqlDatabase* db) {
 }
 
 int MigrateSchema(SqlDatabase* db, int current_version_number) {
+  if (!db) {
+    return 0;
+  }
+
   if (current_version_number == kVersionNumber) {
     return current_version_number;
   }
 
   static constexpr char kQuery[] = "DROP TABLE IF EXISTS annotations";
-  sql::Statement statement = db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  if (!statement.Run()) {
+  std::unique_ptr<sql::Statement> statement =
+      db->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+  if (!statement || !statement->Run()) {
     return 0;
   }
 
   return CreateNewSchema(db);
+}
+
+// Returns sorted `FileSearchResult`s contained in both sorted arrays.
+std::vector<FileSearchResult> FindIntersection(
+    const std::vector<FileSearchResult>& vec1,
+    const std::vector<FileSearchResult>& vec2) {
+  std::vector<FileSearchResult> result;
+
+  auto it1 = vec1.begin();
+  auto it2 = vec2.begin();
+
+  while (it1 != vec1.end() && it2 != vec2.end()) {
+    if (it1->file_path < it2->file_path) {
+      ++it1;
+    } else if (it2->file_path < it1->file_path) {
+      ++it2;
+    } else {
+      result.emplace_back(FileSearchResult(it1->file_path, it1->last_modified,
+                                           it1->relevance + it2->relevance));
+      ++it1;
+      ++it2;
+    }
+  }
+
+  return result;
 }
 
 }  // namespace
@@ -104,6 +138,8 @@ FileSearchResult::FileSearchResult(const base::FilePath& file_path,
 
 FileSearchResult::~FileSearchResult() = default;
 FileSearchResult::FileSearchResult(const FileSearchResult&) = default;
+FileSearchResult& FileSearchResult::operator=(const FileSearchResult&) =
+    default;
 
 AnnotationStorage::AnnotationStorage(
     const base::FilePath& path_to_db,
@@ -154,15 +190,18 @@ void AnnotationStorage::Insert(const ImageInfo& image_info) {
   // clang-format on
 
   for (const auto& annotation : image_info.annotations) {
-    sql::Statement statement =
+    std::unique_ptr<sql::Statement> statement =
         sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+    if (!statement) {
+      return;
+    }
     DVLOG(1) << annotation;
-    statement.BindString(0, annotation);
-    statement.BindString(1, image_info.path.value());
-    statement.BindTime(2, image_info.last_modified);
-    statement.BindInt(3, image_info.is_ignored);
+    statement->BindString(0, annotation);
+    statement->BindString(1, image_info.path.value());
+    statement->BindTime(2, image_info.last_modified);
+    statement->BindInt(3, image_info.is_ignored);
 
-    if (!statement.Run()) {
+    if (!statement->Run()) {
       // TODO(b/260646344): log to UMA instead.
       return;
     }
@@ -176,11 +215,15 @@ void AnnotationStorage::Remove(const base::FilePath& image_path) {
 
   static constexpr char kQuery[] = "DELETE FROM annotations WHERE image_path=?";
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  statement.BindString(0, image_path.value());
+  if (!statement) {
+    return;
+  }
 
-  statement.Run();
+  statement->BindString(0, image_path.value());
+
+  statement->Run();
 }
 
 std::vector<ImageInfo> AnnotationStorage::GetAllAnnotations() {
@@ -194,17 +237,20 @@ std::vector<ImageInfo> AnnotationStorage::GetAllAnnotations() {
           "ORDER BY label";
   // clang-format on
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
+  if (!statement) {
+    return {};
+  }
 
   std::vector<ImageInfo> matched_paths;
-  while (statement.Step()) {
-    const base::FilePath path = base::FilePath(statement.ColumnString(1));
-    const base::Time time = statement.ColumnTime(2);
-    const bool is_ignored = statement.ColumnBool(3);
-    DVLOG(1) << "Select find: " << statement.ColumnString(0) << ", " << path
+  while (statement->Step()) {
+    const base::FilePath path = base::FilePath(statement->ColumnString(1));
+    const base::Time time = statement->ColumnTime(2);
+    const bool is_ignored = statement->ColumnBool(3);
+    DVLOG(1) << "Select find: " << statement->ColumnString(0) << ", " << path
              << ", " << time;
-    matched_paths.push_back({{statement.ColumnString(0)},
+    matched_paths.push_back({{statement->ColumnString(0)},
                              std::move(path),
                              std::move(time),
                              is_ignored});
@@ -227,18 +273,21 @@ std::vector<ImageInfo> AnnotationStorage::FindImagePath(
           "ORDER BY label";
   // clang-format on
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  statement.BindString(0, image_path.value());
+  if (!statement) {
+    return {};
+  }
+  statement->BindString(0, image_path.value());
 
   std::vector<ImageInfo> matched_paths;
-  while (statement.Step()) {
-    const base::FilePath path = base::FilePath(statement.ColumnString(1));
-    const base::Time time = statement.ColumnTime(2);
-    const bool is_ignored = statement.ColumnBool(3);
-    DVLOG(1) << "Select find: " << statement.ColumnString(0) << ", " << path
+  while (statement->Step()) {
+    const base::FilePath path = base::FilePath(statement->ColumnString(1));
+    const base::Time time = statement->ColumnTime(2);
+    const bool is_ignored = statement->ColumnBool(3);
+    DVLOG(1) << "Select find: " << statement->ColumnString(0) << ", " << path
              << ", " << time;
-    matched_paths.push_back({{statement.ColumnString(0)},
+    matched_paths.push_back({{statement->ColumnString(0)},
                              std::move(path),
                              std::move(time),
                              is_ignored});
@@ -247,11 +296,10 @@ std::vector<ImageInfo> AnnotationStorage::FindImagePath(
   return matched_paths;
 }
 
-std::vector<FileSearchResult> AnnotationStorage::Search(
-    const std::u16string& query) {
+std::vector<FileSearchResult> AnnotationStorage::PrefixSearch(
+    const std::u16string& query_term) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
-  DVLOG(1) << "Search";
-  using TokenizedString = ash::string_matching::TokenizedString;
+  DVLOG(1) << "PrefixSearch " << query_term;
 
   // LIKE is 10 times faster than the linear search.
   static constexpr char kQuery[] =
@@ -263,25 +311,28 @@ std::vector<FileSearchResult> AnnotationStorage::Search(
           "ORDER BY image_path";
   // clang-format on
 
-  sql::Statement statement =
+  std::unique_ptr<sql::Statement> statement =
       sql_database_->GetStatementForQuery(SQL_FROM_HERE, kQuery);
-  statement.BindString(0, base::StrCat({"%", base::UTF16ToUTF8(query), "%"}));
+  if (!statement) {
+    return {};
+  }
+  statement->BindString(0, base::StrCat({base::UTF16ToUTF8(query_term), "%"}));
 
   std::vector<FileSearchResult> matched_paths;
-  TokenizedString tokenized_query(query);
-  ash::string_matching::FuzzyTokenizedStringMatch fuzzy_match;
-  while (statement.Step()) {
-    double relevance = fuzzy_match.Relevance(
+  TokenizedString tokenized_query(query_term, Mode::kWords);
+  while (statement->Step()) {
+    double relevance = FuzzyTokenizedStringMatch::TokenSetRatio(
         tokenized_query,
-        TokenizedString(base::UTF8ToUTF16(statement.ColumnString(0))),
-        /*use_weighted_ratio=*/true);
+        TokenizedString(base::UTF8ToUTF16(statement->ColumnString(0)),
+                        Mode::kWords),
+        /*partial=*/false);
     if (relevance < kRelevanceThreshold) {
       continue;
     }
 
-    const base::FilePath path = base::FilePath(statement.ColumnString(1));
-    const base::Time time = statement.ColumnTime(2);
-    DVLOG(1) << "Select: " << statement.ColumnString(0) << ", " << path << ", "
+    const base::FilePath path = base::FilePath(statement->ColumnString(1));
+    const base::Time time = statement->ColumnTime(2);
+    DVLOG(1) << "Select: " << statement->ColumnString(0) << ", " << path << ", "
              << time << " rl: " << relevance;
 
     if (matched_paths.empty() || matched_paths.back().file_path != path) {
@@ -291,6 +342,60 @@ std::vector<FileSearchResult> AnnotationStorage::Search(
     }
   }
   return matched_paths;
+}
+
+std::vector<FileSearchResult> AnnotationStorage::Search(
+    const std::u16string& query,
+    size_t max_num_results) {
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+  if (max_num_results < 1) {
+    return {};
+  }
+
+  TokenizedString tokenized_query(query, Mode::kWords);
+  if (tokenized_query.tokens().empty()) {
+    return {};
+  }
+
+  std::vector<FileSearchResult> results;
+  int normalization_constant = tokenized_query.tokens().size();
+  bool fist_result = true;
+  for (const auto& token : tokenized_query.tokens()) {
+    if (IsStopWord(base::UTF16ToUTF8(token))) {
+      normalization_constant -= 1;
+      continue;
+    }
+
+    std::vector<FileSearchResult> next_result = PrefixSearch(token);
+    if (next_result.empty()) {
+      return {};
+    }
+    results =
+        (fist_result) ? next_result : FindIntersection(results, next_result);
+    fist_result = false;
+  }
+
+  if (results.size() <= max_num_results) {
+    std::sort(results.begin(), results.end(),
+              [](const FileSearchResult& a, const FileSearchResult& b) {
+                return a.relevance > b.relevance;
+              });
+  } else {
+    std::partial_sort(results.begin(), results.begin() + max_num_results,
+                      results.end(),
+                      [](const FileSearchResult& a, const FileSearchResult& b) {
+                        return a.relevance > b.relevance;
+                      });
+    results = std::vector<FileSearchResult>(results.begin(),
+                                            results.begin() + max_num_results);
+  }
+
+  // Normalize to [0, 1].
+  for (auto& result : results) {
+    result.relevance = result.relevance / normalization_constant;
+  }
+
+  return results;
 }
 
 }  // namespace app_list
